+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 <!-- There should always be "Unreleased" section at the beginning. -->
 
 ## Unreleased
+- Add `OnErrorCallback::ignoreError` named constructor
 
 ## 1.0.0 - 2021-05-12
 - Initial implementation